+++ semcore/tag/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [5.27.0] - 2024-02-19
+
+### Changed
+
+- Interactive tags keyboard focused ring style adjusted to same focus ring styles of other components.
+- Interactive tags now get `role=button`.
+
 ## [5.26.0] - 2024-02-13
 
 ### Changed